+++ config/application.rb
@@ -9,10 +9,15 @@ Bundler.require(*Rails.groups)
 module MyFirstProject
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
+
     config.load_defaults 5.1
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
+
+    # Adicionando i18n no projeto
+    config.i18n.default_locale = :"pt-BR"
+    #config.active_record.raise_in_trasactional_callbacks = true
   end
 end